Add files via upload
WAP to accept a dictionary and another container, replace all the value pairs
in the dictionary with the values from the container. If the dictionary len	
is greater then the rest keys will have None value. If only single value is passed
then all the keys of dictionary will have that single value.